# docs/swagger.yaml
basePath: /
definitions:
  apierror.Error:
    properties:
      error:
        type: string
    type: object
  battlelog.CreateMatchInput:
    properties:
      player1_id:
        type: string
      player2_id:
        type: string
      points1:
        maximum: 100
        minimum: 0
        type: integer
      points2:
        maximum: 100
        minimum: 0
        type: integer
      winner_id:
        type: string
    required:
    - player1_id
    - player2_id
    - points1
    - points2
    - winner_id
    type: object
  battlelog.CreateTeamInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  battlelog.MatchOutput:
    properties:
      created_at:
        type: string
      id:
        type: string
      player1_id:
        type: string
      player2_id:
        type: string
      points1:
        type: integer
      points2:
        type: integer
      updated_at:
        type: string
      winner_id:
        type: string
    type: object
  battlelog.TeamOutput:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  battlelog.UpdateMatchInput:
    properties:
      player1_id:
        type: string
      player2_id:
        type: string
      points1:
        maximum: 100
        minimum: 0
        type: integer
      points2:
        maximum: 100
        minimum: 0
        type: integer
      winner_id:
        type: string
    required:
    - player1_id
    - player2_id
    - points1
    - points2
    - winner_id
    type: object
  battlelog.UpdateTeamInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Tournament Records and Metrics Assistant
  title: TRAMA API
  version: "1.0"
paths:
...
  /api/v1/matches:
    get:
      description: List all WH40k 11th matches
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/battlelog.MatchOutput'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: List matches
      tags:
      - matches
    post:
      consumes:
      - application/json
      description: Create a new WH40k 11th match
      parameters:
      - description: Match data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/battlelog.CreateMatchInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/battlelog.MatchOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Create match
      tags:
      - matches
  /api/v1/matches/{id}:
    delete:
      description: Delete a match by ID
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Delete match
      tags:
      - matches
    get:
      description: Get a match by ID
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/battlelog.MatchOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Get match
      tags:
      - matches
    put:
      consumes:
      - application/json
      description: Update a match by ID
      parameters:
      - description: Match ID
        in: path
        name: id
        required: true
        type: string
      - description: Match data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/battlelog.UpdateMatchInput'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Update match
      tags:
      - matches
  /api/v1/teams:
    get:
      description: List all teams
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/battlelog.TeamOutput'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: List teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Create a new team
      parameters:
      - description: Team data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/battlelog.CreateTeamInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/battlelog.TeamOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Create team
      tags:
      - teams
  /api/v1/teams/{id}:
    delete:
      description: Delete a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Delete team
      tags:
      - teams
    get:
      description: Get a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/battlelog.TeamOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Get team
      tags:
      - teams
    put:
      consumes:
      - application/json
      description: Update a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      - description: Team data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/battlelog.UpdateTeamInput'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.Error'
      summary: Update team
      tags:
      - teams
...
swagger: "2.0"